Update StepTimeTredmill View
Updated scaling so regardless of the values for step_time and target_zones will be properly represented in the visualization. Updated visualization css to clearly identify the visualizion and add a key. Updated the positions for targetZones so there is at least 1 on the screen at all times. Updated useStepTime so the initial values are 0 for everything like they were initially to ensure no other files have issues.